backend/requirements.prod.txt
# Production requirements - includes all base requirements plus production-specific packages
-r requirements.txt

# Production WSGI server
gunicorn==21.2.0
gevent==24.2.1

# Performance monitoring
psutil==5.9.5

# Production logging
structlog==23.1.0

# Security
django-ratelimit==4.1.0

# Health checks
django-health-check==3.17.0

# Environment management
python-decouple==3.8

# SAML SSO for Chalmers authentication
# Using pysaml2 instead of python3-saml (better maintained, no xmlsec issues)
pysaml2==7.5.0
djangosaml2==1.9.3

# Privacy policy rendering
markdown==3.5.1

# Optional: Error tracking
sentry-sdk==1.40.0